requirements.txt
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cron-descriptor==1.4.3
dj-database-url==2.2.0
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-heroku==0.3.1
django-jalali==6.2.1
django-jazzmin==3.0.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
locust==2.30.0
MarkupSafe==2.1.5
msgpack==1.0.8
nest-asyncio==1.6.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg==3.2.1
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-git==2018.2.1
python-telegram-bot==21.4
PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
Send2Trash==1.8.3
sentry-sdk==2.10.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
whitenoise==6.7.0
zope.event==5.0
zope.interface==6.4.post2
